自動生成モードをtags-splitに変更し、service系ファイルを分割してframeworkから移動
