+++ example-app/SantokuApp/src/framework/backend/index.ts
@@ -2,4 +2,3 @@ export * from './WithReactQuery';
 export * from './useDefaultOptions';
 export * from './refreshCsrfToken';
 export * from './refreshSession';
-export * from './useBackendService';
