+++ example-app/SantokuApp/src/screens/demo/react-query/CreateTodoDemoScreen.tsx
@@ -2,11 +2,11 @@ import {CompositeScreenProps} from '@react-navigation/native';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
 import {Button} from 'components/button/Button';
 import {useLoadingOverlay} from 'components/overlay';
-import {usePostTodoService} from 'framework/backend';
 import {DemoStackParamList, RootStackParamList} from 'navigation/types';
 import React, {useCallback, useState} from 'react';
 import {StyleSheet, View} from 'react-native';
 import {Input} from 'react-native-elements';
+import {usePostTodoService} from 'service/backend';
 
 import {EditTodoDemoScreen} from './EditTodoDemoScreen';
