+++ example-app/SantokuApp/src/screens/demo/react-query/DisabledQueryDemoScreen.tsx
@@ -1,11 +1,11 @@
 import {CompositeScreenProps} from '@react-navigation/native';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
-import {useListTodoService} from 'framework/backend';
 import {DemoStackParamList, RootStackParamList} from 'navigation/types';
 import React, {useCallback, useMemo, useState} from 'react';
 import {ActivityIndicator, SafeAreaView, ScrollView, StyleSheet, View} from 'react-native';
 import {Text, Button} from 'react-native-elements';
 import {useQueryClient} from 'react-query';
+import {useListTodoService} from 'service/backend';
 
 type Props = CompositeScreenProps<
   NativeStackScreenProps<DemoStackParamList, typeof ScreenName>,
