+++ example-app/SantokuApp/src/screens/demo/react-query/EditTodoDemoScreen.tsx
@@ -2,11 +2,11 @@ import {CompositeScreenProps} from '@react-navigation/native';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
 import {Button} from 'components/button/Button';
 import {useLoadingOverlay} from 'components/overlay';
-import {useDeleteTodoService, useGetTodoService, usePutTodoService} from 'framework/backend';
 import {DemoStackParamList, RootStackParamList} from 'navigation/types';
 import React, {useCallback, useEffect, useState} from 'react';
 import {ActivityIndicator, StyleSheet, View} from 'react-native';
 import {Input, Text} from 'react-native-elements';
+import {useDeleteTodoService, useGetTodoService, usePutTodoService} from 'service/backend';
 
 type Props = CompositeScreenProps<
   NativeStackScreenProps<DemoStackParamList, typeof ScreenName>,
