+++ example-app/SantokuApp/src/screens/demo/react-query/ListTodoDemoScreen.tsx
@@ -1,6 +1,5 @@
 import {CompositeScreenProps} from '@react-navigation/native';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
-import {useListTodoByCursorInfiniteService} from 'framework/backend';
 import {Todo} from 'generated/sandbox/model';
 import {DemoStackParamList, RootStackParamList} from 'navigation/types';
 import React, {useCallback, useMemo, useState} from 'react';
@@ -8,6 +7,7 @@ import {ActivityIndicator, Pressable, RefreshControl, SafeAreaView, StyleSheet, 
 import {Icon, ListItem, Text, FAB, Button} from 'react-native-elements';
 import {FlatList} from 'react-native-gesture-handler';
 import {useQueryClient} from 'react-query';
+import {useListTodoByCursorInfiniteService} from 'service/backend';
 
 import {CreateTodoDemoScreen} from './CreateTodoDemoScreen';
 import {EditTodoDemoScreen} from './EditTodoDemoScreen';
