+++ example-app/SantokuApp/src/service/backend/useTermService.ts
@@ -1,0 +1,8 @@
+import {useGetTerms} from 'generated/backend/terms/terms';
+
+import {BackendServiceQueryOptions, convertQueryResult, getQueryOptions} from './base';
+
+export const useGetTermsService = (options?: BackendServiceQueryOptions) => {
+  const query = useGetTerms({query: getQueryOptions(options)});
+  return convertQueryResult(query);
+};
